Moving regression tests to its own file, to avoid polluting core tests, that can help to understand how dirty fields is working
