+++ tests/test_core.py
@@ -1,0 +1,94 @@
+from decimal import Decimal
+from django.test import TestCase
+from .models import (TestModel, TestModelWithForeignKey, TestModelWithOneToOneField,
+                     SubclassModel, TestModelWithDecimalField)
+
+
+class DirtyFieldsMixinTestCase(TestCase):
+
+    def test_dirty_fields(self):
+        tm = TestModel()
+        # initial state shouldn't be dirty
+        self.assertEqual(tm.get_dirty_fields(), {})
+
+        # changing values should flag them as dirty
+        tm.boolean = False
+        tm.characters = 'testing'
+        self.assertEqual(tm.get_dirty_fields(), {
+            'boolean': True,
+            'characters': ''
+        })
+
+        # resetting them to original values should unflag
+        tm.boolean = True
+        self.assertEqual(tm.get_dirty_fields(), {
+            'characters': ''
+        })
+
+    def test_sweeping(self):
+        tm = TestModel()
+        tm.boolean = False
+        tm.characters = 'testing'
+        self.assertEqual(tm.get_dirty_fields(), {
+            'boolean': True,
+            'characters': ''
+        })
+        tm.save()
+        self.assertEqual(tm.get_dirty_fields(), {})
+
+    def test_relationship_option_for_foreign_key(self):
+        tm1 = TestModel.objects.create()
+        tm2 = TestModel.objects.create()
+        tm = TestModelWithForeignKey.objects.create(fkey=tm1)
+
+        # initial state shouldn't be dirty
+        self.assertEqual(tm.get_dirty_fields(), {})
+
+        # Default dirty check is not taking foreignkeys into account
+        tm.fkey = tm2
+        self.assertEqual(tm.get_dirty_fields(), {})
+
+        # But if we use 'check_relationships' param, then we have to.
+        self.assertEqual(tm.get_dirty_fields(check_relationship=True), {
+            'fkey': tm1.pk
+        })
+
+    def test_relationship_option_for_one_to_one_field(self):
+        tm1 = TestModel.objects.create()
+        tm2 = TestModel.objects.create()
+        tm = TestModelWithOneToOneField.objects.create(o2o=tm1)
+
+        # initial state shouldn't be dirty
+        self.assertEqual(tm.get_dirty_fields(), {})
+
+        # Default dirty check is not taking onetoone fields into account
+        tm.o2o = tm2
+        self.assertEqual(tm.get_dirty_fields(), {})
+
+        # But if we use 'check_relationships' param, then we have to.
+        self.assertEqual(tm.get_dirty_fields(check_relationship=True), {
+            'o2o': tm1.pk
+        })
+
+
+    def test_non_local_fields(self):
+        subclass = SubclassModel.objects.create(characters='foo')
+        subclass.characters = 'spam'
+
+        self.assertTrue(subclass.is_dirty())
+        self.assertDictEqual(subclass.get_dirty_fields(), {'characters': 'foo'})
+
+    def test_decimal_field_correctly_managed(self):
+        # Non regression test case for bug:
+        # https://github.com/smn/django-dirtyfields/issues/4
+        tm = TestModelWithDecimalField.objects.create(decimal_field=Decimal(2.00))
+
+        # initial state shouldn't be dirty
+        self.assertFalse(tm.is_dirty())
+
+        tm.decimal_field = 2.0
+        self.assertFalse(tm.is_dirty())
+
+        tm.decimal_field = u"2.00"
+        self.assertFalse(tm.is_dirty())
+
